requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
redis==4.6.0
aiohttp==3.9.5
openai==1.59.7
python-dotenv==1.0.1
pytest==7.4.3
asyncio==3.4.3